// src/api/types.d.ts
export namespace NQuakeAPI {
    export interface ILeaderboard {
        boardType: string;
        entries: ILeaderboardItem[];
        totalEntries: number;
    }

    export interface ILeaderboardItem {
        userName: string;
        eloRating: number;
        profileIconId: string;
        namePlateId: string;
    }
// ...
}